Move `typeParam` factory to separate file
